Adding check for iOS directory existing before clearing out GPGS code.

+++ source/PluginDev/Assets/GooglePlayGames/Editor/GPGSPostBuild.cs
@@ -39,8 +39,10 @@ namespace GooglePlayGames {
             #if NO_GPGS
             // remove plugin code from generated project
             string pluginDir = pathToBuiltProject + "/Libraries/Plugins/iOS";
-            GPGSUtil.WriteFile(pluginDir + "/GPGSAppController.mm", "// Empty since NO_GPGS is defined\n");
-            return;
+            if(System.IO.Directory.Exists(pluginDir)) {
+                GPGSUtil.WriteFile(pluginDir + "/GPGSAppController.mm", "// Empty since NO_GPGS is defined\n");
+                return;
+            }
             #endif
 
             if (GetBundleId() == null) {
